@@ -4,12 +4,13 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
 
 	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/api/types/system"
 	"github.com/jumppad-labs/jumppad/pkg/clients/container/mocks"
 	imocks "github.com/jumppad-labs/jumppad/pkg/clients/images/mocks"
 	"github.com/jumppad-labs/jumppad/pkg/clients/logger"
@@ -25,7 +26,7 @@ func TestCopyFromContainerCopiesFile(t *testing.T) {
 
 	md := &mocks.Docker{}
 	md.On("ServerVersion", mock.Anything).Return(types.Version{}, nil)
-	md.On("Info", mock.Anything).Return(types.Info{Driver: StorageDriverOverlay2}, nil)
+	md.On("Info", mock.Anything).Return(system.Info{Driver: StorageDriverOverlay2}, nil)
 
 	tmpDir := t.TempDir()
 	tgz := &tar.TarGz{}
@@ -43,7 +44,7 @@ func TestCopyFromContainerCopiesFile(t *testing.T) {
 	mic := &imocks.ImageLog{}
 	md.On("CopyFromContainer", mock.Anything, id, src).Return(
 		io.NopCloser(bytes.NewBuffer(buf.Bytes())),
-		types.ContainerPathStat{},
+		container.PathStat{},
 		nil,
 	)
 
@@ -53,7 +54,7 @@ func TestCopyFromContainerCopiesFile(t *testing.T) {
 	require.NoError(t, err)
 
 	// check the file was written correctly
-	d, err := ioutil.ReadFile(filepath.Join(tmpDir, "output", "file.txt"))
+	d, err := os.ReadFile(filepath.Join(tmpDir, "output", "file.txt"))
 	require.NoError(t, err)
 	require.Equal(t, "test content", string(d))
 }
@@ -64,12 +65,12 @@ func TestCopyFromContainerReturnsErrorOnDockerError(t *testing.T) {
 
 	md := &mocks.Docker{}
 	md.On("ServerVersion", mock.Anything).Return(types.Version{}, nil)
-	md.On("Info", mock.Anything).Return(types.Info{Driver: StorageDriverOverlay2}, nil)
+	md.On("Info", mock.Anything).Return(system.Info{Driver: StorageDriverOverlay2}, nil)
 
 	mic := &imocks.ImageLog{}
 	md.On("CopyFromContainer", mock.Anything, id, src).Return(
 		nil,
-		types.ContainerPathStat{},
+		container.PathStat{},
 		fmt.Errorf("boom"),
 	)
 	dt, _ := NewDockerTasks(md, mic, &tar.TarGz{}, logger.NewTestLogger(t))
