@@ -5,7 +5,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -16,7 +15,6 @@ import (
 	conmocks "github.com/jumppad-labs/jumppad/pkg/clients/connector/mocks"
 	contypes "github.com/jumppad-labs/jumppad/pkg/clients/connector/types"
 	cmocks "github.com/jumppad-labs/jumppad/pkg/clients/container/mocks"
-	"github.com/jumppad-labs/jumppad/pkg/clients/container/types"
 	ctypes "github.com/jumppad-labs/jumppad/pkg/clients/container/types"
 	"github.com/jumppad-labs/jumppad/pkg/clients/k8s"
 	"github.com/jumppad-labs/jumppad/pkg/clients/logger"
@@ -50,9 +48,9 @@ func setupClusterMocks(t *testing.T) (
 	md.On("RemoveContainer", mock.Anything, mock.Anything).Return(nil)
 	md.On("RemoveVolume", mock.Anything).Return(nil)
 	md.On("DetachNetwork", mock.Anything, mock.Anything, mock.Anything).Return(nil)
-	md.On("ListNetworks", mock.Anything).Return([]types.NetworkAttachment{})
+	md.On("ListNetworks", mock.Anything).Return([]ctypes.NetworkAttachment{})
 
-	md.On("EngineInfo").Return(&types.EngineInfo{StorageDriver: "overlay2"})
+	md.On("EngineInfo").Return(&ctypes.EngineInfo{StorageDriver: "overlay2"})
 
 	// set the home folder to a temp folder
 	tmpDir := t.TempDir()
@@ -84,15 +82,42 @@ func setupClusterMocks(t *testing.T) (
 	mk.Mock.On("Apply", mock.Anything, mock.Anything).Return(nil)
 	mk.Mock.On("GetPodLogs", mock.Anything, mock.Anything, mock.Anything).Return(nil, nil)
 
+	rc, err := os.CreateTemp(tmpDir, "root.cert")
+	if err != nil {
+		panic(err)
+	}
+
+	rk, err := os.CreateTemp(tmpDir, "root.key")
+	if err != nil {
+		panic(err)
+	}
+
+	lc, err := os.CreateTemp(tmpDir, "leaf.cert")
+	if err != nil {
+		panic(err)
+	}
+
+	lk, err := os.CreateTemp(tmpDir, "leaf.key")
+	if err != nil {
+		panic(err)
+	}
+
+	bundle := &contypes.CertBundle{
+		RootCertPath: rc.Name(),
+		RootKeyPath:  rk.Name(),
+		LeafCertPath: lc.Name(),
+		LeafKeyPath:  lk.Name(),
+	}
+
 	mc := &conmocks.Connector{}
-	mc.On("GetLocalCertBundle", mock.Anything).Return(&contypes.CertBundle{}, nil)
+	mc.On("GetLocalCertBundle", mock.Anything).Return(bundle, nil)
 	mc.On("GenerateLeafCert",
 		mock.Anything,
 		mock.Anything,
 		mock.Anything,
 		mock.Anything,
 		mock.Anything,
-	).Return(&contypes.CertBundle{}, nil)
+	).Return(bundle, nil)
 
 	// copy the config
 	cc := deepcopy.Copy(clusterConfig).(*Cluster)
@@ -124,7 +149,7 @@ func TestClusterK3SetsEnvironment(t *testing.T) {
 	err := p.Create(context.Background())
 	assert.NoError(t, err)
 
-	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*types.Container)
+	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*ctypes.Container)
 
 	assert.Equal(t, params.Environment["K3S_KUBECONFIG_OUTPUT"], "/output/kubeconfig.yaml")
 	assert.Equal(t, params.Environment["CONTAINERD_HTTP_PROXY"], utils.ImageCacheAddress())
@@ -142,7 +167,7 @@ func TestClusterK3DoesNotSetProxyEnvironmentWithWrongVersion(t *testing.T) {
 	err := p.Create(context.Background())
 	assert.NoError(t, err)
 
-	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*types.Container)
+	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*ctypes.Container)
 
 	assert.Empty(t, params.Environment["CONTAINERD_HTTP_PROXY"])
 }
@@ -156,7 +181,7 @@ func TestClusterK3NoProxyIsSet(t *testing.T) {
 	err := p.Create(context.Background())
 	assert.NoError(t, err)
 
-	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*types.Container)
+	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*ctypes.Container)
 	assert.Equal(t, "test.com,test2.com", params.Environment["CONTAINERD_NO_PROXY"])
 }
 
@@ -177,7 +202,7 @@ func TestClusterK3PullsImage(t *testing.T) {
 
 	err := p.Create(context.Background())
 	assert.NoError(t, err)
-	md.AssertCalled(t, "PullImage", types.Image{Name: "shipyardrun/k3s:v1.27.4"}, false)
+	md.AssertCalled(t, "PullImage", ctypes.Image{Name: "shipyardrun/k3s:v1.27.4"}, false)
 }
 
 func TestClusterK3CreatesNewVolume(t *testing.T) {
@@ -211,7 +236,7 @@ func TestClusterK3CreatesAServer(t *testing.T) {
 	err := p.Create(context.Background())
 	assert.NoError(t, err)
 
-	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*types.Container)
+	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*ctypes.Container)
 
 	// validate the basic details for the server container
 	assert.Contains(t, params.Name, "server")
@@ -265,7 +290,7 @@ func TestClusterK3CreatesAServerWithAdditionalPorts(t *testing.T) {
 	err := p.Create(context.Background())
 	assert.NoError(t, err)
 
-	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*types.Container)
+	params := testutils.GetCalls(&md.Mock, "CreateContainer")[0].Arguments[0].(*ctypes.Container)
 
 	localPort, _ := strconv.Atoi(params.Ports[3].Local)
 	hostPort, _ := strconv.Atoi(params.Ports[3].Host)
@@ -281,7 +306,7 @@ func TestClusterK3sErrorsIfServerNOTStart(t *testing.T) {
 
 	testutils.RemoveOn(&md.Mock, "ContainerLogs")
 	md.On("ContainerLogs", mock.Anything, true, true).Return(
-		ioutil.NopCloser(bytes.NewBufferString("Not running")),
+		io.NopCloser(bytes.NewBufferString("Not running")),
 		nil,
 	)
 
@@ -343,7 +368,7 @@ func TestClusterK3sSetsServerInConfig(t *testing.T) {
 	defer f.Close()
 
 	// check file contains docker ip
-	d, err := ioutil.ReadAll(f)
+	d, err := io.ReadAll(f)
 	assert.NoError(t, err)
 	assert.Contains(t, string(d), "https://"+utils.GetDockerIP())
 }
