@@ -1,7 +1,6 @@
 package config
 
 import (
-	"io/ioutil"
 	"os"
 	"testing"
 
@@ -82,7 +81,7 @@ func CreateTestFile(t *testing.T, contents string) string {
 
 // create a temporary directory
 func createTempDirectory(t *testing.T) string {
-	dir, err := ioutil.TempDir("", "")
+	dir, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("Unable to create temporary directory: %s", err)
 	}
@@ -91,7 +90,7 @@ func createTempDirectory(t *testing.T) string {
 }
 
 func createNamedFile(t *testing.T, dir, name, contents string) string {
-	f, err := ioutil.TempFile(dir, name)
+	f, err := os.CreateTemp(dir, name)
 	if err != nil {
 		t.Fatalf("Error creating temp file %s", err)
 	}
