@@ -52,13 +52,13 @@ func newConnectorCertCmd() *cobra.Command {
 				ca := &crypto.X509{}
 				err := ca.ReadFile(rootCA)
 				if err != nil {
-					return fmt.Errorf("Unable to read root certificate: %s", rootCA)
+					return fmt.Errorf("unable to read root certificate: %s", rootCA)
 				}
 
 				rk := crypto.NewKeyPair()
 				err = rk.Private.ReadFile(rootKey)
 				if err != nil {
-					return fmt.Errorf("Unable to read root certificate: %s", rootKey)
+					return fmt.Errorf("unable to read root certificate: %s", rootKey)
 				}
 
 				k, err := crypto.GenerateKeyPair()
@@ -77,6 +77,10 @@ func newConnectorCertCmd() *cobra.Command {
 				}
 
 				err = lc.WriteFile(path.Join(args[0], "leaf.cert"))
+				if err != nil {
+					return err
+				}
+
 				return nil
 			}
 
