@@ -1,4 +1,4 @@
-// Code generated by mockery v2.46.0. DO NOT EDIT.
+// Code generated by mockery v2.52.3. DO NOT EDIT.
 
 package mocks
 
@@ -131,7 +131,7 @@ func (_m *Connector) GetLocalCertBundle(dir string) (*types.CertBundle, error) {
 	return r0, r1
 }
 
-// IsRunning provides a mock function with given fields:
+// IsRunning provides a mock function with no fields
 func (_m *Connector) IsRunning() bool {
 	ret := _m.Called()
 
@@ -149,7 +149,7 @@ func (_m *Connector) IsRunning() bool {
 	return r0
 }
 
-// ListServices provides a mock function with given fields:
+// ListServices provides a mock function with no fields
 func (_m *Connector) ListServices() ([]*shipyard.Service, error) {
 	ret := _m.Called()
 
@@ -215,7 +215,7 @@ func (_m *Connector) Start(_a0 *types.CertBundle) error {
 	return r0
 }
 
-// Stop provides a mock function with given fields:
+// Stop provides a mock function with no fields
 func (_m *Connector) Stop() error {
 	ret := _m.Called()
 
