@@ -7,6 +7,8 @@ import (
 	"testing"
 
 	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/image"
+	"github.com/docker/docker/api/types/system"
 	"github.com/jumppad-labs/jumppad/pkg/clients/container/mocks"
 	dtypes "github.com/jumppad-labs/jumppad/pkg/clients/container/types"
 	"github.com/jumppad-labs/jumppad/pkg/clients/logger"
@@ -18,7 +20,7 @@ func TestTagImageTagstheImage(t *testing.T) {
 	md := &mocks.Docker{}
 	md.On("ServerVersion", mock.Anything).Return(types.Version{}, nil)
 	md.On("ImageTag", mock.Anything, mock.Anything, mock.Anything).Return(nil)
-	md.On("Info", mock.Anything).Return(types.Info{Driver: StorageDriverOverlay2}, nil)
+	md.On("Info", mock.Anything).Return(system.Info{Driver: StorageDriverOverlay2}, nil)
 
 	dt, err := NewDockerTasks(md, nil, nil, logger.NewTestLogger(t))
 	require.NoError(t, err)
@@ -32,7 +34,7 @@ func TestTagImageTagstheImage(t *testing.T) {
 func TestPushPushestheImageToTheRegistryWithoutAuth(t *testing.T) {
 	md := &mocks.Docker{}
 	md.On("ServerVersion", mock.Anything).Return(types.Version{}, nil)
-	md.On("Info", mock.Anything).Return(types.Info{Driver: StorageDriverOverlay2}, nil)
+	md.On("Info", mock.Anything).Return(system.Info{Driver: StorageDriverOverlay2}, nil)
 	md.On("ImagePush", mock.Anything, mock.Anything, mock.Anything).Return(io.NopCloser(&bytes.Buffer{}), nil)
 
 	dt, err := NewDockerTasks(md, nil, nil, logger.NewTestLogger(t))
@@ -45,7 +47,7 @@ func TestPushPushestheImageToTheRegistryWithoutAuth(t *testing.T) {
 
 	// ensure auth is set with a domain only
 	args := md.Calls[2].Arguments
-	auth := args.Get(2).(types.ImagePushOptions).RegistryAuth
+	auth := args.Get(2).(image.PushOptions).RegistryAuth
 	authString, _ := base64.StdEncoding.DecodeString(auth)
 
 	require.Contains(t, string(authString), `"serveraddress":"docker.io"`)
@@ -56,7 +58,7 @@ func TestPushPushestheImageToTheRegistryWithoutAuth(t *testing.T) {
 func TestPushPushestheImageToTheRegistryWithAuth(t *testing.T) {
 	md := &mocks.Docker{}
 	md.On("ServerVersion", mock.Anything).Return(types.Version{}, nil)
-	md.On("Info", mock.Anything).Return(types.Info{Driver: StorageDriverOverlay2}, nil)
+	md.On("Info", mock.Anything).Return(system.Info{Driver: StorageDriverOverlay2}, nil)
 	md.On("ImagePush", mock.Anything, mock.Anything, mock.Anything).Return(io.NopCloser(&bytes.Buffer{}), nil)
 
 	dt, err := NewDockerTasks(md, nil, nil, logger.NewTestLogger(t))
@@ -69,7 +71,7 @@ func TestPushPushestheImageToTheRegistryWithAuth(t *testing.T) {
 
 	// ensure auth is not set
 	args := md.Calls[2].Arguments
-	auth := args.Get(2).(types.ImagePushOptions).RegistryAuth
+	auth := args.Get(2).(image.PushOptions).RegistryAuth
 	authString, _ := base64.StdEncoding.DecodeString(auth)
 
 	require.Contains(t, string(authString), "user")
