@@ -1,4 +1,4 @@
-// Code generated by mockery v2.46.0. DO NOT EDIT.
+// Code generated by mockery v2.52.3. DO NOT EDIT.
 
 package mocks
 
@@ -326,7 +326,7 @@ func (_m *ContainerTasks) DetachNetwork(network string, containerid string) erro
 	return r0
 }
 
-// EngineInfo provides a mock function with given fields:
+// EngineInfo provides a mock function with no fields
 func (_m *ContainerTasks) EngineInfo() *types.EngineInfo {
 	ret := _m.Called()
 
