@@ -1,14 +1,24 @@
-// Code generated by mockery v2.46.0. DO NOT EDIT.
+// Code generated by mockery v2.52.3. DO NOT EDIT.
 
 package mocks
 
 import (
+	checkpoint "github.com/docker/docker/api/types/checkpoint"
+	client "github.com/docker/docker/client"
+
+	common "github.com/docker/docker/api/types/common"
+
 	context "context"
+
+	image "github.com/docker/docker/api/types/image"
+
 	io "io"
 
 	mock "github.com/stretchr/testify/mock"
 
 	network "github.com/docker/docker/api/types/network"
+
+	system "github.com/docker/docker/api/types/system"
 
 	types "github.com/docker/docker/api/types"
 
@@ -25,7 +35,7 @@ type Docker struct {
 }
 
 // CheckpointCreate provides a mock function with given fields: ctx, _a1, options
-func (_m *Docker) CheckpointCreate(ctx context.Context, _a1 string, options types.CheckpointCreateOptions) error {
+func (_m *Docker) CheckpointCreate(ctx context.Context, _a1 string, options checkpoint.CreateOptions) error {
 	ret := _m.Called(ctx, _a1, options)
 
 	if len(ret) == 0 {
@@ -33,7 +43,7 @@ func (_m *Docker) CheckpointCreate(ctx context.Context, _a1 string, options type
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.CheckpointCreateOptions) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, checkpoint.CreateOptions) error); ok {
 		r0 = rf(ctx, _a1, options)
 	} else {
 		r0 = ret.Error(0)
@@ -43,27 +53,27 @@ func (_m *Docker) CheckpointCreate(ctx context.Context, _a1 string, options type
 }
 
 // CheckpointList provides a mock function with given fields: ctx, _a1, options
-func (_m *Docker) CheckpointList(ctx context.Context, _a1 string, options types.CheckpointListOptions) ([]types.Checkpoint, error) {
+func (_m *Docker) CheckpointList(ctx context.Context, _a1 string, options checkpoint.ListOptions) ([]checkpoint.Summary, error) {
 	ret := _m.Called(ctx, _a1, options)
 
 	if len(ret) == 0 {
 		panic("no return value specified for CheckpointList")
 	}
 
-	var r0 []types.Checkpoint
+	var r0 []checkpoint.Summary
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.CheckpointListOptions) ([]types.Checkpoint, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, checkpoint.ListOptions) ([]checkpoint.Summary, error)); ok {
 		return rf(ctx, _a1, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.CheckpointListOptions) []types.Checkpoint); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, checkpoint.ListOptions) []checkpoint.Summary); ok {
 		r0 = rf(ctx, _a1, options)
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).([]types.Checkpoint)
+			r0 = ret.Get(0).([]checkpoint.Summary)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.CheckpointListOptions) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, checkpoint.ListOptions) error); ok {
 		r1 = rf(ctx, _a1, options)
 	} else {
 		r1 = ret.Error(1)
@@ -101,7 +111,7 @@ func (_m *Docker) ContainerCreate(ctx context.Context, config *typescontainer.Co
 }
 
 // ContainerExecAttach provides a mock function with given fields: ctx, execID, config
-func (_m *Docker) ContainerExecAttach(ctx context.Context, execID string, config types.ExecStartCheck) (types.HijackedResponse, error) {
+func (_m *Docker) ContainerExecAttach(ctx context.Context, execID string, config typescontainer.ExecStartOptions) (types.HijackedResponse, error) {
 	ret := _m.Called(ctx, execID, config)
 
 	if len(ret) == 0 {
@@ -110,16 +120,16 @@ func (_m *Docker) ContainerExecAttach(ctx context.Context, execID string, config
 
 	var r0 types.HijackedResponse
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ExecStartCheck) (types.HijackedResponse, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.ExecStartOptions) (types.HijackedResponse, error)); ok {
 		return rf(ctx, execID, config)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ExecStartCheck) types.HijackedResponse); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.ExecStartOptions) types.HijackedResponse); ok {
 		r0 = rf(ctx, execID, config)
 	} else {
 		r0 = ret.Get(0).(types.HijackedResponse)
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.ExecStartCheck) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, typescontainer.ExecStartOptions) error); ok {
 		r1 = rf(ctx, execID, config)
 	} else {
 		r1 = ret.Error(1)
@@ -129,25 +139,25 @@ func (_m *Docker) ContainerExecAttach(ctx context.Context, execID string, config
 }
 
 // ContainerExecCreate provides a mock function with given fields: ctx, _a1, config
-func (_m *Docker) ContainerExecCreate(ctx context.Context, _a1 string, config types.ExecConfig) (types.IDResponse, error) {
+func (_m *Docker) ContainerExecCreate(ctx context.Context, _a1 string, config typescontainer.ExecOptions) (common.IDResponse, error) {
 	ret := _m.Called(ctx, _a1, config)
 
 	if len(ret) == 0 {
 		panic("no return value specified for ContainerExecCreate")
 	}
 
-	var r0 types.IDResponse
+	var r0 common.IDResponse
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ExecConfig) (types.IDResponse, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.ExecOptions) (common.IDResponse, error)); ok {
 		return rf(ctx, _a1, config)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ExecConfig) types.IDResponse); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.ExecOptions) common.IDResponse); ok {
 		r0 = rf(ctx, _a1, config)
 	} else {
-		r0 = ret.Get(0).(types.IDResponse)
+		r0 = ret.Get(0).(common.IDResponse)
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.ExecConfig) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, typescontainer.ExecOptions) error); ok {
 		r1 = rf(ctx, _a1, config)
 	} else {
 		r1 = ret.Error(1)
@@ -157,22 +167,22 @@ func (_m *Docker) ContainerExecCreate(ctx context.Context, _a1 string, config ty
 }
 
 // ContainerExecInspect provides a mock function with given fields: ctx, execID
-func (_m *Docker) ContainerExecInspect(ctx context.Context, execID string) (types.ContainerExecInspect, error) {
+func (_m *Docker) ContainerExecInspect(ctx context.Context, execID string) (typescontainer.ExecInspect, error) {
 	ret := _m.Called(ctx, execID)
 
 	if len(ret) == 0 {
 		panic("no return value specified for ContainerExecInspect")
 	}
 
-	var r0 types.ContainerExecInspect
+	var r0 typescontainer.ExecInspect
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string) (types.ContainerExecInspect, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string) (typescontainer.ExecInspect, error)); ok {
 		return rf(ctx, execID)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string) types.ContainerExecInspect); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string) typescontainer.ExecInspect); ok {
 		r0 = rf(ctx, execID)
 	} else {
-		r0 = ret.Get(0).(types.ContainerExecInspect)
+		r0 = ret.Get(0).(typescontainer.ExecInspect)
 	}
 
 	if rf, ok := ret.Get(1).(func(context.Context, string) error); ok {
@@ -185,7 +195,7 @@ func (_m *Docker) ContainerExecInspect(ctx context.Context, execID string) (type
 }
 
 // ContainerExecResize provides a mock function with given fields: ctx, execID, config
-func (_m *Docker) ContainerExecResize(ctx context.Context, execID string, config types.ResizeOptions) error {
+func (_m *Docker) ContainerExecResize(ctx context.Context, execID string, config typescontainer.ResizeOptions) error {
 	ret := _m.Called(ctx, execID, config)
 
 	if len(ret) == 0 {
@@ -193,7 +203,7 @@ func (_m *Docker) ContainerExecResize(ctx context.Context, execID string, config
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ResizeOptions) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.ResizeOptions) error); ok {
 		r0 = rf(ctx, execID, config)
 	} else {
 		r0 = ret.Error(0)
@@ -203,7 +213,7 @@ func (_m *Docker) ContainerExecResize(ctx context.Context, execID string, config
 }
 
 // ContainerExecStart provides a mock function with given fields: ctx, execID, config
-func (_m *Docker) ContainerExecStart(ctx context.Context, execID string, config types.ExecStartCheck) error {
+func (_m *Docker) ContainerExecStart(ctx context.Context, execID string, config typescontainer.ExecStartOptions) error {
 	ret := _m.Called(ctx, execID, config)
 
 	if len(ret) == 0 {
@@ -211,7 +221,7 @@ func (_m *Docker) ContainerExecStart(ctx context.Context, execID string, config 
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ExecStartCheck) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.ExecStartOptions) error); ok {
 		r0 = rf(ctx, execID, config)
 	} else {
 		r0 = ret.Error(0)
@@ -221,22 +231,22 @@ func (_m *Docker) ContainerExecStart(ctx context.Context, execID string, config 
 }
 
 // ContainerInspect provides a mock function with given fields: ctx, containerID
-func (_m *Docker) ContainerInspect(ctx context.Context, containerID string) (types.ContainerJSON, error) {
+func (_m *Docker) ContainerInspect(ctx context.Context, containerID string) (typescontainer.InspectResponse, error) {
 	ret := _m.Called(ctx, containerID)
 
 	if len(ret) == 0 {
 		panic("no return value specified for ContainerInspect")
 	}
 
-	var r0 types.ContainerJSON
+	var r0 typescontainer.InspectResponse
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string) (types.ContainerJSON, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string) (typescontainer.InspectResponse, error)); ok {
 		return rf(ctx, containerID)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string) types.ContainerJSON); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string) typescontainer.InspectResponse); ok {
 		r0 = rf(ctx, containerID)
 	} else {
-		r0 = ret.Get(0).(types.ContainerJSON)
+		r0 = ret.Get(0).(typescontainer.InspectResponse)
 	}
 
 	if rf, ok := ret.Get(1).(func(context.Context, string) error); ok {
@@ -249,27 +259,27 @@ func (_m *Docker) ContainerInspect(ctx context.Context, containerID string) (typ
 }
 
 // ContainerList provides a mock function with given fields: ctx, options
-func (_m *Docker) ContainerList(ctx context.Context, options types.ContainerListOptions) ([]types.Container, error) {
+func (_m *Docker) ContainerList(ctx context.Context, options typescontainer.ListOptions) ([]typescontainer.Summary, error) {
 	ret := _m.Called(ctx, options)
 
 	if len(ret) == 0 {
 		panic("no return value specified for ContainerList")
 	}
 
-	var r0 []types.Container
+	var r0 []typescontainer.Summary
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, types.ContainerListOptions) ([]types.Container, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, typescontainer.ListOptions) ([]typescontainer.Summary, error)); ok {
 		return rf(ctx, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, types.ContainerListOptions) []types.Container); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, typescontainer.ListOptions) []typescontainer.Summary); ok {
 		r0 = rf(ctx, options)
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).([]types.Container)
+			r0 = ret.Get(0).([]typescontainer.Summary)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, types.ContainerListOptions) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, typescontainer.ListOptions) error); ok {
 		r1 = rf(ctx, options)
 	} else {
 		r1 = ret.Error(1)
@@ -279,7 +289,7 @@ func (_m *Docker) ContainerList(ctx context.Context, options types.ContainerList
 }
 
 // ContainerLogs provides a mock function with given fields: ctx, _a1, options
-func (_m *Docker) ContainerLogs(ctx context.Context, _a1 string, options types.ContainerLogsOptions) (io.ReadCloser, error) {
+func (_m *Docker) ContainerLogs(ctx context.Context, _a1 string, options typescontainer.LogsOptions) (io.ReadCloser, error) {
 	ret := _m.Called(ctx, _a1, options)
 
 	if len(ret) == 0 {
@@ -288,10 +298,10 @@ func (_m *Docker) ContainerLogs(ctx context.Context, _a1 string, options types.C
 
 	var r0 io.ReadCloser
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ContainerLogsOptions) (io.ReadCloser, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.LogsOptions) (io.ReadCloser, error)); ok {
 		return rf(ctx, _a1, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ContainerLogsOptions) io.ReadCloser); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.LogsOptions) io.ReadCloser); ok {
 		r0 = rf(ctx, _a1, options)
 	} else {
 		if ret.Get(0) != nil {
@@ -299,7 +309,7 @@ func (_m *Docker) ContainerLogs(ctx context.Context, _a1 string, options types.C
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.ContainerLogsOptions) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, typescontainer.LogsOptions) error); ok {
 		r1 = rf(ctx, _a1, options)
 	} else {
 		r1 = ret.Error(1)
@@ -309,7 +319,7 @@ func (_m *Docker) ContainerLogs(ctx context.Context, _a1 string, options types.C
 }
 
 // ContainerRemove provides a mock function with given fields: ctx, containerID, options
-func (_m *Docker) ContainerRemove(ctx context.Context, containerID string, options types.ContainerRemoveOptions) error {
+func (_m *Docker) ContainerRemove(ctx context.Context, containerID string, options typescontainer.RemoveOptions) error {
 	ret := _m.Called(ctx, containerID, options)
 
 	if len(ret) == 0 {
@@ -317,7 +327,7 @@ func (_m *Docker) ContainerRemove(ctx context.Context, containerID string, optio
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ContainerRemoveOptions) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.RemoveOptions) error); ok {
 		r0 = rf(ctx, containerID, options)
 	} else {
 		r0 = ret.Error(0)
@@ -327,7 +337,7 @@ func (_m *Docker) ContainerRemove(ctx context.Context, containerID string, optio
 }
 
 // ContainerStart provides a mock function with given fields: _a0, _a1, _a2
-func (_m *Docker) ContainerStart(_a0 context.Context, _a1 string, _a2 types.ContainerStartOptions) error {
+func (_m *Docker) ContainerStart(_a0 context.Context, _a1 string, _a2 typescontainer.StartOptions) error {
 	ret := _m.Called(_a0, _a1, _a2)
 
 	if len(ret) == 0 {
@@ -335,7 +345,7 @@ func (_m *Docker) ContainerStart(_a0 context.Context, _a1 string, _a2 types.Cont
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ContainerStartOptions) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, typescontainer.StartOptions) error); ok {
 		r0 = rf(_a0, _a1, _a2)
 	} else {
 		r0 = ret.Error(0)
@@ -363,7 +373,7 @@ func (_m *Docker) ContainerStop(ctx context.Context, containerID string, options
 }
 
 // CopyFromContainer provides a mock function with given fields: ctx, containerID, srcPath
-func (_m *Docker) CopyFromContainer(ctx context.Context, containerID string, srcPath string) (io.ReadCloser, types.ContainerPathStat, error) {
+func (_m *Docker) CopyFromContainer(ctx context.Context, containerID string, srcPath string) (io.ReadCloser, typescontainer.PathStat, error) {
 	ret := _m.Called(ctx, containerID, srcPath)
 
 	if len(ret) == 0 {
@@ -371,9 +381,9 @@ func (_m *Docker) CopyFromContainer(ctx context.Context, containerID string, src
 	}
 
 	var r0 io.ReadCloser
-	var r1 types.ContainerPathStat
+	var r1 typescontainer.PathStat
 	var r2 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, string) (io.ReadCloser, types.ContainerPathStat, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, string) (io.ReadCloser, typescontainer.PathStat, error)); ok {
 		return rf(ctx, containerID, srcPath)
 	}
 	if rf, ok := ret.Get(0).(func(context.Context, string, string) io.ReadCloser); ok {
@@ -384,10 +394,10 @@ func (_m *Docker) CopyFromContainer(ctx context.Context, containerID string, src
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, string) types.ContainerPathStat); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, string) typescontainer.PathStat); ok {
 		r1 = rf(ctx, containerID, srcPath)
 	} else {
-		r1 = ret.Get(1).(types.ContainerPathStat)
+		r1 = ret.Get(1).(typescontainer.PathStat)
 	}
 
 	if rf, ok := ret.Get(2).(func(context.Context, string, string) error); ok {
@@ -400,7 +410,7 @@ func (_m *Docker) CopyFromContainer(ctx context.Context, containerID string, src
 }
 
 // CopyToContainer provides a mock function with given fields: ctx, _a1, path, content, options
-func (_m *Docker) CopyToContainer(ctx context.Context, _a1 string, path string, content io.Reader, options types.CopyToContainerOptions) error {
+func (_m *Docker) CopyToContainer(ctx context.Context, _a1 string, path string, content io.Reader, options typescontainer.CopyToContainerOptions) error {
 	ret := _m.Called(ctx, _a1, path, content, options)
 
 	if len(ret) == 0 {
@@ -408,7 +418,7 @@ func (_m *Docker) CopyToContainer(ctx context.Context, _a1 string, path string, 
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, string, io.Reader, types.CopyToContainerOptions) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, string, io.Reader, typescontainer.CopyToContainerOptions) error); ok {
 		r0 = rf(ctx, _a1, path, content, options)
 	} else {
 		r0 = ret.Error(0)
@@ -446,27 +456,27 @@ func (_m *Docker) ImageBuild(ctx context.Context, buildContext io.Reader, option
 }
 
 // ImageList provides a mock function with given fields: ctx, options
-func (_m *Docker) ImageList(ctx context.Context, options types.ImageListOptions) ([]types.ImageSummary, error) {
+func (_m *Docker) ImageList(ctx context.Context, options image.ListOptions) ([]image.Summary, error) {
 	ret := _m.Called(ctx, options)
 
 	if len(ret) == 0 {
 		panic("no return value specified for ImageList")
 	}
 
-	var r0 []types.ImageSummary
+	var r0 []image.Summary
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, types.ImageListOptions) ([]types.ImageSummary, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, image.ListOptions) ([]image.Summary, error)); ok {
 		return rf(ctx, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, types.ImageListOptions) []types.ImageSummary); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, image.ListOptions) []image.Summary); ok {
 		r0 = rf(ctx, options)
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).([]types.ImageSummary)
+			r0 = ret.Get(0).([]image.Summary)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, types.ImageListOptions) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, image.ListOptions) error); ok {
 		r1 = rf(ctx, options)
 	} else {
 		r1 = ret.Error(1)
@@ -476,7 +486,7 @@ func (_m *Docker) ImageList(ctx context.Context, options types.ImageListOptions)
 }
 
 // ImagePull provides a mock function with given fields: ctx, refStr, options
-func (_m *Docker) ImagePull(ctx context.Context, refStr string, options types.ImagePullOptions) (io.ReadCloser, error) {
+func (_m *Docker) ImagePull(ctx context.Context, refStr string, options image.PullOptions) (io.ReadCloser, error) {
 	ret := _m.Called(ctx, refStr, options)
 
 	if len(ret) == 0 {
@@ -485,10 +495,10 @@ func (_m *Docker) ImagePull(ctx context.Context, refStr string, options types.Im
 
 	var r0 io.ReadCloser
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ImagePullOptions) (io.ReadCloser, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, image.PullOptions) (io.ReadCloser, error)); ok {
 		return rf(ctx, refStr, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ImagePullOptions) io.ReadCloser); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, image.PullOptions) io.ReadCloser); ok {
 		r0 = rf(ctx, refStr, options)
 	} else {
 		if ret.Get(0) != nil {
@@ -496,7 +506,7 @@ func (_m *Docker) ImagePull(ctx context.Context, refStr string, options types.Im
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.ImagePullOptions) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, image.PullOptions) error); ok {
 		r1 = rf(ctx, refStr, options)
 	} else {
 		r1 = ret.Error(1)
@@ -505,9 +515,13 @@ func (_m *Docker) ImagePull(ctx context.Context, refStr string, options types.Im
 	return r0, r1
 }
 
-// ImagePush provides a mock function with given fields: ctx, image, options
-func (_m *Docker) ImagePush(ctx context.Context, image string, options types.ImagePushOptions) (io.ReadCloser, error) {
-	ret := _m.Called(ctx, image, options)
+// ImagePush provides a mock function with given fields: ctx, _a1, options
+func (_m *Docker) ImagePush(ctx context.Context, _a1 string, options image.PushOptions) (io.ReadCloser, error) {
+	ret := _m.Called(ctx, _a1, options)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ImagePush")
+	}
 
 	if len(ret) == 0 {
 		panic("no return value specified for ImagePush")
@@ -515,19 +529,19 @@ func (_m *Docker) ImagePush(ctx context.Context, image string, options types.Ima
 
 	var r0 io.ReadCloser
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ImagePushOptions) (io.ReadCloser, error)); ok {
-		return rf(ctx, image, options)
+	if rf, ok := ret.Get(0).(func(context.Context, string, image.PushOptions) (io.ReadCloser, error)); ok {
+		return rf(ctx, _a1, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ImagePushOptions) io.ReadCloser); ok {
-		r0 = rf(ctx, image, options)
+	if rf, ok := ret.Get(0).(func(context.Context, string, image.PushOptions) io.ReadCloser); ok {
+		r0 = rf(ctx, _a1, options)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(io.ReadCloser)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.ImagePushOptions) error); ok {
-		r1 = rf(ctx, image, options)
+	if rf, ok := ret.Get(1).(func(context.Context, string, image.PushOptions) error); ok {
+		r1 = rf(ctx, _a1, options)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -536,27 +550,27 @@ func (_m *Docker) ImagePush(ctx context.Context, image string, options types.Ima
 }
 
 // ImageRemove provides a mock function with given fields: ctx, imageID, options
-func (_m *Docker) ImageRemove(ctx context.Context, imageID string, options types.ImageRemoveOptions) ([]types.ImageDeleteResponseItem, error) {
+func (_m *Docker) ImageRemove(ctx context.Context, imageID string, options image.RemoveOptions) ([]image.DeleteResponse, error) {
 	ret := _m.Called(ctx, imageID, options)
 
 	if len(ret) == 0 {
 		panic("no return value specified for ImageRemove")
 	}
 
-	var r0 []types.ImageDeleteResponseItem
+	var r0 []image.DeleteResponse
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ImageRemoveOptions) ([]types.ImageDeleteResponseItem, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, image.RemoveOptions) ([]image.DeleteResponse, error)); ok {
 		return rf(ctx, imageID, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.ImageRemoveOptions) []types.ImageDeleteResponseItem); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, image.RemoveOptions) []image.DeleteResponse); ok {
 		r0 = rf(ctx, imageID, options)
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).([]types.ImageDeleteResponseItem)
+			r0 = ret.Get(0).([]image.DeleteResponse)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.ImageRemoveOptions) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, image.RemoveOptions) error); ok {
 		r1 = rf(ctx, imageID, options)
 	} else {
 		r1 = ret.Error(1)
@@ -565,9 +579,20 @@ func (_m *Docker) ImageRemove(ctx context.Context, imageID string, options types
 	return r0, r1
 }
 
-// ImageSave provides a mock function with given fields: ctx, imageIDs
-func (_m *Docker) ImageSave(ctx context.Context, imageIDs []string) (io.ReadCloser, error) {
-	ret := _m.Called(ctx, imageIDs)
+// ImageSave provides a mock function with given fields: ctx, imageIDs, saveOpts
+func (_m *Docker) ImageSave(ctx context.Context, imageIDs []string, saveOpts ...client.ImageSaveOption) (io.ReadCloser, error) {
+	_va := make([]interface{}, len(saveOpts))
+	for _i := range saveOpts {
+		_va[_i] = saveOpts[_i]
+	}
+	var _ca []interface{}
+	_ca = append(_ca, ctx, imageIDs)
+	_ca = append(_ca, _va...)
+	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ImageSave")
+	}
 
 	if len(ret) == 0 {
 		panic("no return value specified for ImageSave")
@@ -575,19 +600,19 @@ func (_m *Docker) ImageSave(ctx context.Context, imageIDs []string) (io.ReadClos
 
 	var r0 io.ReadCloser
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, []string) (io.ReadCloser, error)); ok {
-		return rf(ctx, imageIDs)
+	if rf, ok := ret.Get(0).(func(context.Context, []string, ...client.ImageSaveOption) (io.ReadCloser, error)); ok {
+		return rf(ctx, imageIDs, saveOpts...)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, []string) io.ReadCloser); ok {
-		r0 = rf(ctx, imageIDs)
+	if rf, ok := ret.Get(0).(func(context.Context, []string, ...client.ImageSaveOption) io.ReadCloser); ok {
+		r0 = rf(ctx, imageIDs, saveOpts...)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(io.ReadCloser)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, []string) error); ok {
-		r1 = rf(ctx, imageIDs)
+	if rf, ok := ret.Get(1).(func(context.Context, []string, ...client.ImageSaveOption) error); ok {
+		r1 = rf(ctx, imageIDs, saveOpts...)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -614,22 +639,22 @@ func (_m *Docker) ImageTag(ctx context.Context, source string, target string) er
 }
 
 // Info provides a mock function with given fields: ctx
-func (_m *Docker) Info(ctx context.Context) (types.Info, error) {
+func (_m *Docker) Info(ctx context.Context) (system.Info, error) {
 	ret := _m.Called(ctx)
 
 	if len(ret) == 0 {
 		panic("no return value specified for Info")
 	}
 
-	var r0 types.Info
+	var r0 system.Info
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context) (types.Info, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context) (system.Info, error)); ok {
 		return rf(ctx)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context) types.Info); ok {
+	if rf, ok := ret.Get(0).(func(context.Context) system.Info); ok {
 		r0 = rf(ctx)
 	} else {
-		r0 = ret.Get(0).(types.Info)
+		r0 = ret.Get(0).(system.Info)
 	}
 
 	if rf, ok := ret.Get(1).(func(context.Context) error); ok {
@@ -660,25 +685,25 @@ func (_m *Docker) NetworkConnect(ctx context.Context, networkID string, containe
 }
 
 // NetworkCreate provides a mock function with given fields: ctx, name, options
-func (_m *Docker) NetworkCreate(ctx context.Context, name string, options types.NetworkCreate) (types.NetworkCreateResponse, error) {
+func (_m *Docker) NetworkCreate(ctx context.Context, name string, options network.CreateOptions) (network.CreateResponse, error) {
 	ret := _m.Called(ctx, name, options)
 
 	if len(ret) == 0 {
 		panic("no return value specified for NetworkCreate")
 	}
 
-	var r0 types.NetworkCreateResponse
+	var r0 network.CreateResponse
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.NetworkCreate) (types.NetworkCreateResponse, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, network.CreateOptions) (network.CreateResponse, error)); ok {
 		return rf(ctx, name, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.NetworkCreate) types.NetworkCreateResponse); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, network.CreateOptions) network.CreateResponse); ok {
 		r0 = rf(ctx, name, options)
 	} else {
-		r0 = ret.Get(0).(types.NetworkCreateResponse)
+		r0 = ret.Get(0).(network.CreateResponse)
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.NetworkCreate) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, network.CreateOptions) error); ok {
 		r1 = rf(ctx, name, options)
 	} else {
 		r1 = ret.Error(1)
@@ -706,25 +731,25 @@ func (_m *Docker) NetworkDisconnect(ctx context.Context, networkID string, conta
 }
 
 // NetworkInspect provides a mock function with given fields: ctx, networkID, options
-func (_m *Docker) NetworkInspect(ctx context.Context, networkID string, options types.NetworkInspectOptions) (types.NetworkResource, error) {
+func (_m *Docker) NetworkInspect(ctx context.Context, networkID string, options network.InspectOptions) (network.Inspect, error) {
 	ret := _m.Called(ctx, networkID, options)
 
 	if len(ret) == 0 {
 		panic("no return value specified for NetworkInspect")
 	}
 
-	var r0 types.NetworkResource
+	var r0 network.Inspect
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.NetworkInspectOptions) (types.NetworkResource, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, network.InspectOptions) (network.Inspect, error)); ok {
 		return rf(ctx, networkID, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, string, types.NetworkInspectOptions) types.NetworkResource); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, network.InspectOptions) network.Inspect); ok {
 		r0 = rf(ctx, networkID, options)
 	} else {
-		r0 = ret.Get(0).(types.NetworkResource)
+		r0 = ret.Get(0).(network.Inspect)
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, string, types.NetworkInspectOptions) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, network.InspectOptions) error); ok {
 		r1 = rf(ctx, networkID, options)
 	} else {
 		r1 = ret.Error(1)
@@ -734,27 +759,27 @@ func (_m *Docker) NetworkInspect(ctx context.Context, networkID string, options 
 }
 
 // NetworkList provides a mock function with given fields: ctx, options
-func (_m *Docker) NetworkList(ctx context.Context, options types.NetworkListOptions) ([]types.NetworkResource, error) {
+func (_m *Docker) NetworkList(ctx context.Context, options network.ListOptions) ([]network.Inspect, error) {
 	ret := _m.Called(ctx, options)
 
 	if len(ret) == 0 {
 		panic("no return value specified for NetworkList")
 	}
 
-	var r0 []types.NetworkResource
+	var r0 []network.Inspect
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, types.NetworkListOptions) ([]types.NetworkResource, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, network.ListOptions) ([]network.Inspect, error)); ok {
 		return rf(ctx, options)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, types.NetworkListOptions) []types.NetworkResource); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, network.ListOptions) []network.Inspect); ok {
 		r0 = rf(ctx, options)
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).([]types.NetworkResource)
+			r0 = ret.Get(0).([]network.Inspect)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, types.NetworkListOptions) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, network.ListOptions) error); ok {
 		r1 = rf(ctx, options)
 	} else {
 		r1 = ret.Error(1)
