@@ -62,7 +62,7 @@ func TestParsesOutput(t *testing.T) {
 	err := p.Create(context.Background())
 	require.NoError(t, err)
 
-	require.True(t, e.Output.AsValueMap()["FOO"].AsString() == "BAR")
+	require.True(t, e.Output["FOO"] == "BAR")
 }
 
 func TestDeletesOutput(t *testing.T) {
