@@ -5,14 +5,12 @@ import (
 	"fmt"
 	"net"
 
-	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/filters"
 	"github.com/docker/docker/api/types/network"
 	htypes "github.com/jumppad-labs/hclconfig/types"
 	"github.com/jumppad-labs/jumppad/pkg/clients"
 	"github.com/jumppad-labs/jumppad/pkg/clients/container"
 	sdk "github.com/jumppad-labs/plugin-sdk"
-	"golang.org/x/xerrors"
 )
 
 var _ sdk.Provider = &Provider{}
@@ -124,7 +122,7 @@ func (p *Provider) Destroy(ctx context.Context, force bool) error {
 	// check network exists if so remove
 	ids, err := p.Lookup()
 	if err != nil {
-		return xerrors.Errorf("Unable to list networks: %w", err)
+		return fmt.Errorf("unable to list networks: %w", err)
 	}
 
 	if len(ids) == 1 {
@@ -168,10 +166,10 @@ func (p *Provider) Changed() (bool, error) {
 }
 
 func (p *Provider) createWithDriver(driver string) error {
-	opts := types.NetworkCreate{
-		CheckDuplicate: true,
-		Driver:         driver,
-		EnableIPv6:     p.config.EnableIPv6,
+	opts := network.CreateOptions{
+		// CheckDuplicate: true,
+		Driver:     driver,
+		EnableIPv6: &p.config.EnableIPv6,
 		IPAM: &network.IPAM{
 			Driver: "default",
 			Config: []network.IPAMConfig{
@@ -192,10 +190,10 @@ func (p *Provider) createWithDriver(driver string) error {
 	return err
 }
 
-func (p *Provider) getNetworks(name string) ([]types.NetworkResource, error) {
+func (p *Provider) getNetworks(name string) ([]network.Summary, error) {
 	args := filters.NewArgs()
 	args.Add("name", name)
-	return p.client.NetworkList(context.Background(), types.NetworkListOptions{Filters: args})
+	return p.client.NetworkList(context.Background(), network.ListOptions{Filters: args})
 }
 
 func (p *Provider) getHostIPs() ([]net.IP, error) {
