@@ -135,9 +135,7 @@ func (n *NomadCluster) Process() error {
 			}
 
 			// the network name is set
-			for x, net := range state.Networks {
-				n.Networks[x] = net
-			}
+			copy(n.Networks, state.Networks)
 		}
 	}
 
