@@ -6,7 +6,6 @@ import (
 	"log"
 	"os"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/jumppad-labs/hclconfig"
@@ -16,6 +15,7 @@ import (
 	"github.com/jumppad-labs/jumppad/pkg/config/mocks"
 	"github.com/jumppad-labs/jumppad/pkg/config/resources/cache"
 	"github.com/jumppad-labs/jumppad/pkg/config/resources/container"
+	"github.com/jumppad-labs/jumppad/pkg/config/resources/network"
 	"github.com/jumppad-labs/jumppad/pkg/jumppad/constants"
 	"github.com/jumppad-labs/jumppad/pkg/utils"
 	"github.com/jumppad-labs/jumppad/testutils"
@@ -23,8 +23,6 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
-
-var lock = sync.Mutex{}
 
 func setupTests(t *testing.T, returnVals map[string]error) (*EngineImpl, *mocks.Providers) {
 	return setupTestsBase(t, returnVals, "")
@@ -52,7 +50,7 @@ func setupTestsBase(t *testing.T, returnVals map[string]error, state string) (*E
 	return e, pm
 }
 
-func testLoadState(t *testing.T, e *EngineImpl) *hclconfig.Config {
+func testLoadState(t *testing.T) *hclconfig.Config {
 	c, err := config.LoadState()
 	require.NoError(t, err)
 
@@ -76,17 +74,20 @@ func TestApplyWithSingleFile(t *testing.T) {
 	_, err := e.Apply(context.Background(), "../../examples/single_file/container.hcl")
 	require.NoError(t, err)
 
-	require.Len(t, e.config.Resources, 7) // 6 resources in the file plus the image cache
+	require.Len(t, e.config.Resources, 8) // 6 resources in the file plus the image cache and default network
 
 	// Check the provider was called for each resource
 	require.ElementsMatch(t,
 		[]string{
-			"default",
-			"onprem",
-			"default",
-			"consul_config",
-			"port_range",
-			"version",
+			"default",       // image cache
+			"jumppad",       // default network
+			"version",       // variable
+			"port_range",    // variable
+			"consul_config", // template
+			"onprem",        // network
+			"default",       // image cache = refresh after adding network
+			"consul",        // container
+			"consul_addr",   // output
 		},
 		[]string{
 			getMetaFromMock(mp, 0).Name,
@@ -95,6 +96,9 @@ func TestApplyWithSingleFile(t *testing.T) {
 			getMetaFromMock(mp, 3).Name,
 			getMetaFromMock(mp, 4).Name,
 			getMetaFromMock(mp, 5).Name,
+			getMetaFromMock(mp, 6).Name,
+			getMetaFromMock(mp, 7).Name,
+			getMetaFromMock(mp, 8).Name,
 		},
 	)
 }
@@ -177,7 +181,9 @@ func TestApplyAddsNetworksToImageCache(t *testing.T) {
 	require.NoError(t, err)
 
 	// network should be added as a dependency
-	require.Equal(t, "resource.network.onprem", r.GetDependencies()[0])
+	require.Len(t, r.GetDependencies(), 2)
+	require.Equal(t, network.DefaultNetworkID, r.GetDependencies()[0])
+	require.Equal(t, "resource.network.onprem", r.GetDependencies()[1])
 }
 
 func TestApplyAddsCustomRegistriesToImageCache(t *testing.T) {
@@ -197,21 +203,31 @@ func TestApplyAddsCustomRegistriesToImageCache(t *testing.T) {
 	require.Len(t, r.(*cache.ImageCache).Registries, 2)
 }
 
+func TestApplyAddsDefaultNetwork(t *testing.T) {
+	e, _ := setupTests(t, nil)
+
+	_, err := e.Apply(context.Background(), "../../examples/default_network/main.hcl")
+	require.NoError(t, err)
+
+	dc := e.ResourceCountForType(network.TypeNetwork)
+	require.Equal(t, 1, dc)
+}
+
 func TestApplyWithSingleFileAndVariables(t *testing.T) {
 	e, mp := setupTests(t, nil)
 
 	_, err := e.ApplyWithVariables(context.Background(), "../../examples/single_file/container.hcl", nil, "../../examples/single_file/default.vars")
 	require.NoError(t, err)
-	require.Len(t, e.config.Resources, 7) // 6 resources in the file plus the image cache
+	require.Len(t, e.config.Resources, 8) // 6 resources in the file plus the image cache and default network
 
 	// then the container should be created
-	require.Equal(t, "consul", getMetaFromMock(mp, 6).Name)
+	require.Equal(t, "consul", getMetaFromMock(mp, 7).Name)
 
 	// finally the provider for the image cache should be updated
-	require.Equal(t, "consul_addr", getMetaFromMock(mp, 7).Name)
+	require.Equal(t, "consul_addr", getMetaFromMock(mp, 8).Name)
 
 	// check the variable has overridden the image
-	cont := getResourceFromMock(mp, 6).(*container.Container)
+	cont := getResourceFromMock(mp, 7).(*container.Container)
 	require.Equal(t, "consul:1.8.1", cont.Image.Name)
 }
 
@@ -229,8 +245,8 @@ func TestApplyCallsProviderCreateForEachProvider(t *testing.T) {
 	testAssertMethodCalled(t, mp, "Refresh", 1)
 
 	// the state should also contain 11 resources
-	sf := testLoadState(t, e)
-	require.Equal(t, 11, sf.ResourceCount())
+	sf := testLoadState(t)
+	require.Equal(t, 12, sf.ResourceCount())
 }
 
 func TestApplyDoesNotCallsProviderCreateWhenInState(t *testing.T) {
@@ -243,8 +259,8 @@ func TestApplyDoesNotCallsProviderCreateWhenInState(t *testing.T) {
 	testAssertMethodCalled(t, mp, "Create", 5)
 
 	// the state should also contain 7 resources
-	sf := testLoadState(t, e)
-	require.Equal(t, 7, sf.ResourceCount())
+	sf := testLoadState(t)
+	require.Equal(t, 8, sf.ResourceCount())
 }
 
 func TestApplyRemovesItemsInStateWhenNotInFiles(t *testing.T) {
@@ -260,8 +276,8 @@ func TestApplyRemovesItemsInStateWhenNotInFiles(t *testing.T) {
 	testAssertMethodCalled(t, mp, "Destroy", 2)
 
 	// the state should also contain 7 resources
-	sf := testLoadState(t, e)
-	require.Equal(t, 7, sf.ResourceCount())
+	sf := testLoadState(t)
+	require.Equal(t, 8, sf.ResourceCount())
 }
 
 func TestApplyNotCallsProviderCreateForDisabledResources(t *testing.T) {
@@ -272,13 +288,13 @@ func TestApplyNotCallsProviderCreateForDisabledResources(t *testing.T) {
 	require.NoError(t, err)
 
 	// should have call create for non disabled resources
-	testAssertMethodCalled(t, mp, "Create", 3) // ImageCache is always created
+	testAssertMethodCalled(t, mp, "Create", 4) // ImageCache and default network are always created
 
 	// disabled resources should still be added to the state
-	sf := testLoadState(t, e)
+	sf := testLoadState(t)
 
-	// should contain 3 from the config plus the image cache
-	require.Equal(t, 4, sf.ResourceCount())
+	// should contain 3 from the config plus the image cache and default network
+	require.Equal(t, 5, sf.ResourceCount())
 
 	// the resource should be in the state but there should be no status
 	r, err := sf.FindResource("resource.container.consul_disabled")
@@ -294,15 +310,15 @@ func TestApplyShouldNotAddDuplicateDisabledResources(t *testing.T) {
 	require.NoError(t, err)
 
 	// should have call create for each provider
-	testAssertMethodCalled(t, mp, "Create", 1) // ImageCache is always created
+	testAssertMethodCalled(t, mp, "Create", 1) // ImageCache and default network are always created
 
 	// disabled resources should still be added to the state
-	sf := testLoadState(t, e)
+	sf := testLoadState(t)
 
-	// should contain 3 from the config plus the image cache
+	// should contain 3 from the config plus the image cache and default network
 	// should not duplicate the disabled container as this already exists in the
 	// state
-	require.Equal(t, 4, sf.ResourceCount())
+	require.Equal(t, 5, sf.ResourceCount())
 
 	// the status should be set to disabled
 	r, err := sf.FindResource("resource.container.consul_disabled")
@@ -316,12 +332,12 @@ func TestApplySetsCreatedStatusForEachResource(t *testing.T) {
 	_, err := e.Apply(context.Background(), "../../examples/single_file/container.hcl")
 	require.NoError(t, err)
 
-	require.Equal(t, 7, e.config.ResourceCount())
+	require.Equal(t, 8, e.config.ResourceCount())
 
 	// should only call create and destroy for the cache as this is pending update
-	testAssertMethodCalled(t, mp, "Create", 7) // ImageCache is always created
+	testAssertMethodCalled(t, mp, "Create", 8) // ImageCache and default network are always created
 
-	sf := testLoadState(t, e)
+	sf := testLoadState(t)
 
 	r, err := sf.FindResource("resource.container.consul")
 	require.NoError(t, err)
@@ -345,10 +361,10 @@ func TestApplyCallsProviderGenerateErrorStopsExecution(t *testing.T) {
 	// should have call create for each provider
 	// there are two top level config items, template and network
 	// network will fail
-	// ImageCache should always be created
-	testAssertMethodCalled(t, mp, "Create", 5)
+	// ImageCache and default network should always be created
+	testAssertMethodCalled(t, mp, "Create", 6)
 
-	sf := testLoadState(t, e)
+	sf := testLoadState(t)
 
 	// should set failed status for network
 	r, err := sf.FindResource("resource.network.onprem")
@@ -373,7 +389,7 @@ func TestApplyCallsProviderDestroyAndCreateForFailedResources(t *testing.T) {
 
 	// should have call create for each provider
 	testAssertMethodCalled(t, mp, "Destroy", 1)
-	testAssertMethodCalled(t, mp, "Create", 7) // ImageCache is always created
+	testAssertMethodCalled(t, mp, "Create", 8) // ImageCache and default network are always created
 }
 
 func TestApplyCallsProviderDestroyForTaintedResources(t *testing.T) {
@@ -384,7 +400,7 @@ func TestApplyCallsProviderDestroyForTaintedResources(t *testing.T) {
 
 	// should have call create for each provider
 	testAssertMethodCalled(t, mp, "Destroy", 1)
-	testAssertMethodCalled(t, mp, "Create", 7) // ImageCache is always created
+	testAssertMethodCalled(t, mp, "Create", 8) // ImageCache and default network are always created
 }
 
 func TestApplyCallsProviderDestroyForDisabledResources(t *testing.T) {
@@ -405,7 +421,7 @@ func TestApplyCallsProviderDestroyForDisabledResources(t *testing.T) {
 
 	// should have call create for each provider
 	testAssertMethodCalled(t, mp, "Destroy", 1, r)
-	testAssertMethodCalled(t, mp, "Create", 0) // ImageCache is always created
+	testAssertMethodCalled(t, mp, "Create", 0) // ImageCache and default network are always created
 }
 
 func TestApplyCallsProviderRefreshForCreatedResources(t *testing.T) {
@@ -438,7 +454,7 @@ func TestDestroyCallsProviderDestroyForEachProvider(t *testing.T) {
 
 	// should have call create for each provider
 	// and once for the image cache
-	testAssertMethodCalled(t, mp, "Destroy", 4)
+	testAssertMethodCalled(t, mp, "Destroy", 5)
 
 	// state should be removed
 	require.NoFileExists(t, utils.StatePath())
@@ -486,8 +502,8 @@ func TestDestroyNotCallsProviderDestroyForResourcesDisabled(t *testing.T) {
 	require.NoError(t, err)
 
 	// should have call create for each provider
-	testAssertMethodCalled(t, mp, "Destroy", 2)
-	testAssertMethodCalled(t, mp, "Create", 0) // ImageCache is always created
+	testAssertMethodCalled(t, mp, "Destroy", 3)
+	testAssertMethodCalled(t, mp, "Create", 0) // ImageCache and default network are always created
 
 	// state should be removed
 	require.NoFileExists(t, utils.StatePath())
@@ -500,7 +516,7 @@ func TestDestroyCallsProviderGenerateErrorStopsExecution(t *testing.T) {
 	require.Error(t, err)
 
 	// should have call destroy for each provider
-	testAssertMethodCalled(t, mp, "Destroy", 2)
+	testAssertMethodCalled(t, mp, "Destroy", 3)
 
 	// state should not be removed
 	require.FileExists(t, utils.StatePath())
@@ -725,6 +741,16 @@ var existingState = `
       },
       "subnet": "10.15.0.0/16"
   },
+	{
+      "meta": {
+        "name": "jumppad",
+        "properties": {
+          "status": "created"
+        },
+        "type": "network"
+      },
+      "subnet": "10.0.10.0/24"
+  },
   {
       "meta": {
         "name": "default",
@@ -762,6 +788,16 @@ var existingState = `
 var singleFileState = `
 {
   "resources": [
+	{
+      "meta": {
+        "name": "jumppad",
+        "properties": {
+          "status": "created"
+        },
+        "type": "network"
+      },
+      "subnet": "10.0.10.0/24"
+  },
   {
       "meta": {
         "name": "onprem",
@@ -818,6 +854,16 @@ var singleFileState = `
 var complexState = `
 {
   "resources": [
+	{
+      "meta": {
+        "name": "jumppad",
+        "properties": {
+          "status": "created"
+        },
+        "type": "network"
+      },
+      "subnet": "10.0.10.0/24"
+  },
   {
       "meta": {
         "name": "cloud",
@@ -865,6 +911,16 @@ var disabledState = `
 {
   "blueprint": null,
   "resources": [
+	{
+      "meta": {
+        "name": "jumppad",
+        "properties": {
+          "status": "created"
+        },
+        "type": "network"
+      },
+      "subnet": "10.0.10.0/24"
+  },
   {
       "meta": {
         "name": "default",
@@ -899,6 +955,16 @@ var disabledAndCreatedState = `
 {
   "blueprint": null,
   "resources": [
+	{
+      "meta": {
+        "name": "jumppad",
+        "properties": {
+          "status": "created"
+        },
+        "type": "network"
+      },
+      "subnet": "10.0.10.0/24"
+  },
   {
       "meta": {
         "name": "default",
