@@ -2,9 +2,9 @@ package utils
 
 import "fmt"
 
-var InvalidBlueprintURIError = fmt.Errorf("inavlid blueprint URI")
-var NameExceedsMaxLengthError = fmt.Errorf("name exceeds the max length of 128 characters")
-var NameContainsInvalidCharactersError = fmt.Errorf("name contains invalid characters characters must be either a-z, A-Z, 0-9, -, _")
+var ErrInvalidBlueprintURI = fmt.Errorf("inavlid blueprint URI")
+var ErrNameExceedsMaxLength = fmt.Errorf("name exceeds the max length of 128 characters")
+var ErrNameContainsInvalidCharacters = fmt.Errorf("name contains invalid characters characters must be either a-z, A-Z, 0-9, -, _")
 
 // ImageVolumeName is the name of the volume which stores the images for clusters
 const ImageVolumeName string = "images"
